requirements.txt
sentence-transformers
pandas
scikit-learn
google-genai
numpy
tqdm
nest_asyncio
joblib
langchain-openai
matplotlib
seaborn
scipy
kaggle
pyarrow
python-dotenv
